add SKLearnCollector, SKLearnAggregator SVM Aggregatpr
