+++ examples/sklearn-svm/job_configs/sklearn_svm_base/app/config/config_fed_server.json
@@ -18,8 +18,15 @@
       "args": {}
     },
     {
-      "id": "aggregator",
-      "path": "svm_model_aggregator.SVMAggregator",
+      "id": "collector",
+      "path": "nvflare.app_opt.sklearn.sklearn_collector.SKLearnCollector",
+      "args": {
+        "aggregator_id" : "svm_aggregator"
+      }
+    },
+    {
+      "id": "svm_aggregator",
+      "path": "svm_aggregator.SVMAggregator",
       "args": {}
     }
   ],
@@ -32,7 +39,7 @@
         "num_rounds": 2,
         "start_round": 0,
         "wait_time_after_min_received": 0,
-        "aggregator_id": "aggregator",
+        "aggregator_id": "collector",
         "persistor_id": "persistor",
         "shareable_generator_id": "shareable_generator",
         "train_task_name": "train",
