+++ examples/sklearn-svm/job_configs/sklearn_svm_base/app/custom/svm_aggregator.py
@@ -1,0 +1,56 @@
+# Copyright (c) 2021-2022, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# We will move to this app_common when it gets matured
+from typing import Dict
+import numpy as np
+from sklearn.svm import SVC
+
+from nvflare.apis.dxo import DataKind
+from nvflare.app_opt.sklearn.sklearn_aggegator import SKLearnAggregator
+
+
+class SVMAggregator(SKLearnAggregator):
+
+    def __init__(self):
+        super().__init__(data_kind=DataKind.WEIGHTS)
+
+    def get_model_params(self, data: dict):
+        return {"support_x": data["support_x"],
+                "support_y": data["support_y"]}
+
+    def aggregate(self, current_round: int, data: Dict[str, dict]) -> dict:
+        # Fist round, collect all support vectors and
+        # perform one round of SVM to produce global model
+        support_x = []
+        support_y = []
+        for client in self.accumulator:
+            client_model = self.accumulator[client]
+            support_x.append(client_model["support_x"])
+            support_y.append(client_model["support_y"])
+        global_x = np.concatenate(support_x)
+        global_y = np.concatenate(support_y)
+        svm_global = SVC(kernel="rbf")
+        svm_global.fit(global_x, global_y)
+
+        index = svm_global.support_
+        support_x = global_x[index]
+        support_y = global_y[index]
+
+        return {"support_x": support_x, "support_y": support_y}
+
+    def reset(self) -> None:
+        # Reset accumulator for next round,
+        # # but not the center and count, which will be used as the starting point of the next round
+        self.accumulator = {}
